+++ Changelog.md
@@ -6,6 +6,17 @@ The file was started with Version `0.4`.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.6] unreleased
+
+### Added
+
+* A robustified version of the Riemannian Levenberg Marquardt algorithm
+
+### Changed
+
+* (breaking change) renamed `CoordinateVectorialType` to `CoefficientVectorialType` to have a
+  consistent naming that anything with respect to a basis is called “coefficients”
+
 ## [0.5.32] January 15, 2026
 
 ### Fixed